refactor: 공통으로 사용하는 타입들 따로 분리
