+++ src/domain/types.ts
@@ -1,0 +1,11 @@
+export interface CardBoxType {
+  cardCompany?: string;
+  cardNumber?: string;
+  cardHolder?: string;
+  expiredDate?: string;
+  type?: 'small' | 'big';
+  color?: string;
+  nickname?: string;
+}
+
+export type CardCompanyType = { cardCompany: string, color: string };
