+++ src/hooks/useCard.ts
@@ -1,10 +1,10 @@
 import { useCallback, useState } from 'react';
-import { CardBoxType } from '../components/CardBox';
+import { CardBoxType } from '../domain/types';
 
 export default function useCard(initialState: CardBoxType) {
   const [cardState, setState] = useState(initialState);
 
-  const setCardState = useCallback((newState) => {
+  const setCardState = useCallback((newState: CardBoxType) => {
     setState({
       ...newState,
     });
