+++ src/pages/CardList.tsx
@@ -1,6 +1,7 @@
 import { NavLink } from 'react-router-dom';
 import { cardRepository } from '../repository';
 import { CardBox, PageTitle } from '../components';
+import { CardBoxType } from '../domain/types';
 
 export default function CardList() {
   const cardList = cardRepository.getItem();
@@ -8,10 +9,10 @@ export default function CardList() {
   return (
     <div className="app flex-column-center">
       <PageTitle title="보유 카드"></PageTitle>
-      {cardList.map((item) => (
+      {cardList.map((item: CardBoxType) => (
         <div className="flex-column-center" key={item.cardNumber}>
           <CardBox {...item} />
-          {item.nickName && <span className="card-nickname">{item.nickName}</span>}
+          {item.nickname && <span className="card-nickname">{item.nickname}</span>}
         </div>
       ))}
       <div className="card-box">
